Move statically generated data to /static
