+++ core/data/index.js
@@ -1,8 +1,5 @@
 // @flow
-import CARDS from './CARDS';
-import CARD_LIST from './CARD_LIST';
-import NAMES from './NAMES';
 import ALIASES from './ALIASES';
-import DICTIONARY from './DICTIONARY';
+import generateDictionary from './generateDictionary';
 
-export { CARDS, CARD_LIST, NAMES, ALIASES, DICTIONARY };
+export { generateDictionary, ALIASES };
