+++ core/data/static/CARD_IDS.js
