+++ core/data/static/CARD_LIST.js
@@ -1,5 +1,5 @@
 // @flow
-import type { Card } from '../types';
+import type { Card } from '../../types';
 import CARDS from './CARDS';
 
 const CARD_LIST: Card[] = Object.keys(CARDS).map(key => CARDS[key]);
