+++ core/data/static/NAMES.js
