+++ core/data/static/README.md
@@ -1,0 +1,4 @@
+All files in `core/data/static` are for data generated from the statically imported `gwent-data-release`.
+These are used for testing purpose and for the website.
+
+The extension does not use them. Instead it fetches the updated data from a Web API.
